requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
numpy==1.20.0
opencv-python>=4.1.1
Pillow>=7.1.2
